Fix executing code in live segfault issue
I suspect the in-memory llvm ir (as string) was
badly passed to llvm and some chars where not
correctly parsed.

Strangely putting the string in a file and passing
that file was fine.

The issue was fixed by first converting the llvm ir
to cow str and then passing it to llvm engine
as bytes